Initial refactoring of Block Gallery blocks
