+++ src/blocks/gallery-carousel/icons.js
