+++ src/blocks/gallery-carousel/index.js
@@ -9,8 +9,8 @@ import filter from 'lodash/filter';
  */
 import './styles/style.scss';
 import './styles/editor.scss';
-import icons from './components/icons';
-import edit from './components/edit';
+import icons from './icons';
+import edit from './edit';
 import { BackgroundStyles } from '../../components/block-gallery/background/';
 import { GlobalAttributes, GlobalTransforms, GlobalClasses } from '../../components/block-gallery/global/';
 
@@ -26,7 +26,7 @@ const { getColorClassName, RichText } = wp.editor;
  */
 const name = 'gallery-carousel';
 
-const title = __( 'Carousel Gallery' );
+const title = __( 'Carousel' );
 
 const icon = icons.carousel;
 
@@ -307,125 +307,6 @@ const settings = {
 			</div>
 		);
 	},
-
-	deprecated: [
-		{
-			attributes: {
-				...blockAttributes,
-			},
-
-			save( { attributes, className } ) {
-
-				const {
-					autoPlay,
-					autoPlaySpeed,
-					captionColor,
-					customCaptionColor,
-					draggable,
-					gridSize,
-					gutter,
-					gutterMobile,
-					height,
-					images,
-					pageDots,
-					prevNextButtons,
-					primaryCaption,
-				} = attributes;
-
-				const wrapperClasses = classnames(
-					'is-cropped',
-					...GlobalClasses( attributes ), {
-						[ `has-horizontal-gutter` ] : gutter > 0,
-					}
-				);
-
-				const wrapperStyles = {
-					...BackgroundStyles( attributes ),
-				};
-
-				const flickityClasses = classnames(
-					`has-carousel`,
-					`has-carousel-${ gridSize }`, {}
-				);
-
-				const flickityStyles = {
-					height: height ? height + 'px' : undefined,
-				};
-
-				const figureClasses = classnames(
-					'blockgallery--figure', {
-						[ `has-margin-left-${ gutter }` ] : gutter > 0,
-						[ `has-margin-left-mobile-${ gutterMobile }` ] : gutterMobile > 0,
-						[ `has-margin-right-${ gutter }` ] : gutter > 0,
-						[ `has-margin-right-mobile-${ gutterMobile }` ] : gutterMobile > 0,
-					}
-				);
-
-				const flickityOptions = {
-					autoPlay: autoPlay && autoPlaySpeed ? autoPlaySpeed : false,
-					draggable: draggable,
-					pageDots: pageDots,
-					prevNextButtons: prevNextButtons,
-					wrapAround: true,
-					arrowShape: {
-						x0: 10,
-						x1: 60, y1: 50,
-						x2: 65, y2: 45,
-						x3: 20
-					},
-				};
-
-				const captionColorClass = getColorClassName( 'color', captionColor );
-
-				const captionClasses = classnames(
-					'blockgallery--caption',
-					'blockgallery--primary-caption',
-					captionColorClass, {
-						'has-caption-color': captionColorClass,
-
-					}
-				);
-
-				const captionStyles = {
-					color: captionColorClass ? undefined : customCaptionColor,
-				};
-
-				// Return early if there are no images.
-				if ( images.length <= 0 ) {
-					return;
-				}
-
-				return (
-					<div className={ className }>
-						<div
-							className={ wrapperClasses }
-							style={ wrapperStyles }
-						>
-							<div
-								className={ flickityClasses }
-								style={ flickityStyles }
-								data-flickity={ JSON.stringify( flickityOptions ) }
-							>
-								{ images.map( ( image ) => {
-
-									const img = <img src={ image.url } alt={ image.alt } data-id={ image.id } data-link={ image.link } className={ image.id ? `wp-image-${ image.id }` : null } />;
-
-									return (
-										<li key={ image.id || image.url } className="blockgallery--item">
-											<figure className={ figureClasses }>
-												{ img }
-											</figure>
-										</li>
-									);
-								} ) }
-							</div>
-						</div>
-						{ ! RichText.isEmpty( primaryCaption ) && <RichText.Content tagName="figcaption" className={ captionClasses } value={ primaryCaption } style={ captionStyles }/> }
-					</div>
-				);
-			},
-		}
-	],
 }
 
 export { name, title, icon, settings };
