+++ src/blocks/gallery-carousel/inspector.js
@@ -1,11 +1,11 @@
 /**
  * Internal dependencies
  */
-import { title } from '../'
-import ResponsiveTabsControl from '../../../components/block-gallery/responsive-tabs-control';
-import SizeControl from '../../../components/block-gallery/size-control';
-import SliderPanel from '../../../components/block-gallery/slider-panel';
-import { BackgroundPanel } from '../../../components/block-gallery/background';
+import { title } from './'
+import ResponsiveTabsControl from '../../components/block-gallery/responsive-tabs-control';
+import SizeControl from '../../components/block-gallery/size-control';
+import SliderPanel from '../../components/block-gallery/slider-panel';
+import { BackgroundPanel } from '../../components/block-gallery/background';
 
 /**
  * WordPress dependencies
