+++ src/blocks/gallery-masonry/icons.js
