+++ src/blocks/gallery-stacked/icons.js
