+++ src/blocks/gallery-stacked/styles/editor.scss
@@ -1,10 +1,4 @@
 .editor-block-list__block[data-type="coblocks/gallery-stacked"] {
-	// Set a min-height for the "Upload an Image" trigger for this gallery.
-	.components-blockgallery-gallery-item__upload {
-		min-height: 100px;
-		width: 100% !important;
-	}
-
 	// Fix issue where selected images are not wider than the viewport.
 	.blockgallery--item:not(.blockgallery--item-uploader) .blockgallery--figure {
 		display: inline-block;
@@ -13,5 +7,10 @@
 
 	.has-fullwidth-images .blockgallery--item:not(.blockgallery--item-uploader) .blockgallery--figure {
 		width: 100% !important;
+	}
+
+	.is-appender {
+		// Remove unnecessary margin applied to the core appender.
+		margin-bottom: 0;
 	}
 }
