+++ src/components/block-gallery/global/index.js
@@ -3,7 +3,6 @@
  */
 import GlobalAttributes from './attributes';
 import GlobalClasses from './classes';
-import GlobalToolbar from './toolbar';
 import GlobalTransforms from './transforms';
 import GlobalStyles from './styles';
 
@@ -13,7 +12,6 @@ import GlobalStyles from './styles';
 export {
 	GlobalAttributes,
 	GlobalClasses,
-	GlobalToolbar,
 	GlobalTransforms,
 	GlobalStyles,
 };
