--- src/components/block-gallery/lightbox/attributes.js
@@ -1,16 +1,0 @@
-/**
- * Set the attributes for this component
- * @type {Object}
- */
-const LightboxAttributes = {
-	lightbox: {
-		type: 'boolean',
-		default: false,
-	},
-	lightboxStyle: {
-		type: 'string',
-		default: 'light',
-	},
-};
-
-export default LightboxAttributes;
