--- src/components/block-gallery/lightbox/classes.js
@@ -1,11 +1,0 @@
-/**
- * CSS classes
- */
-function LightboxClasses( attributes ) {
-	return [
-		{ 'has-lightbox': attributes.lightbox },
-		{ [ `has-lightbox-style-${ attributes.lightboxStyle }` ] : attributes.lightbox && attributes.lightboxStyle },
-	];
-}
-
-export default LightboxClasses;
