--- src/components/block-gallery/lightbox/index.js
@@ -1,17 +1,0 @@
-/**
- * Internal dependencies
- */
-import LightboxAttributes from './attributes';
-import LightboxClasses from './classes';
-import LightboxTransforms from './transforms';
-import LightboxControl from './control';
-
-/**
- * Export
- */
-export {
-	LightboxAttributes,
-	LightboxClasses,
-	LightboxTransforms,
-	LightboxControl,
-};
