--- src/components/block-gallery/lightbox/options.js
@@ -1,14 +1,0 @@
-/**
- * WordPress dependencies
- */
-const { __ } = wp.i18n;
-
-/**
- * Style options.
- */
-const lightboxStyleOptions = [
-	{ value: 'light', label: __( 'Light' ) },
-	{ value: 'dark', label: __( 'Dark' ) },
-];
-
-export default lightboxStyleOptions;
