--- src/components/block-gallery/lightbox/transforms.js
@@ -1,15 +1,0 @@
-/**
- * Set the attributes for the component transformations
- * @type {Object}
- */
-function LightboxTransforms( props ) {
-
-	const transforms = {
-		lightbox: props.lightbox,
-		lightboxStyle: props.lightboxStyle,
-	};
-
-	return transforms;
-}
-
-export default LightboxTransforms;
